refactor: 데이터 패칭 로직 수정
useIssues에서 이뤄지는 작업들을 리덕스의 thunk와 useInfiniteScroll에서 처리하도록 변경 했습니다.

기존은 리스트페이지에서 마지막리스트 요소를 만나면 toNextPage 액션을 디스패치 => 리듀서에서 페이지를 1증가하고 로딩상태를 "notLoaded"로 변경 =>
useIssues 훅의 useEffect안에서 "notLoaded"상태를 감지하면 fetchIssues thunk 액션을 디스패 치 => 데이터 패칭 =>
fullfill 되면 스토어의 이슈리스트에 이슈추가 라는 복잡한 로직을 가지고있었습니다.

바뀐 로직은 리스트페이지에서 마지막리스트 요소를 만나면 fetchissues thunk 액션을 디스패치 =>
thunk 내부에서 페이지를 가져와 데이터를 패칭 => fullfill 되면 이슈리스트에 이슈추가하고 페이지를 1증가
