+++ src/pages/IssueListPage.tsx
@@ -2,30 +2,12 @@ import Advertisement from 'components/common/Advertisement'
 import IssueInfo from 'components/common/IssueInfo'
 import SpinnerBubble from 'components/common/SpinnerBubble'
 import { ITEM_PER_ADS } from 'constants/github'
-import useIssues from 'hooks/useIssues'
-import { useEffect, useRef } from 'react'
+import useInfiniteScroll from 'hooks/useInfiniteScroll'
 import { StyledLink } from 'styles/GlobalStyles'
 import { LastElment, List, ListItem } from 'styles/issue/issueListStyles'
 
 const IssueListPage = () => {
-  const [issues, toNextPage, isLoading] = useIssues()
-  const lastElRef = useRef<HTMLLIElement>(null)
-  const containerRef = useRef<HTMLUListElement>(null)
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting) toNextPage()
-      },
-      {
-        root: containerRef.current,
-      },
-    )
-
-    observer.observe(lastElRef.current!)
-
-    return () => observer.disconnect()
-  }, [toNextPage])
+  const { containerRef, lastElRef, issues, isLoading } = useInfiniteScroll()
 
   return (
     <List ref={containerRef}>
