+++ src/redux/issueSlice.ts
@@ -2,22 +2,25 @@ import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
 import github from 'apis/github'
 import type { Issues } from 'types/issue'
 
+import { RootState } from './store'
+
 interface IssuesState {
   issues: Issues
   currentPage: number
-  loadingState: 'loaded' | 'notLoaded' | 'loading'
+  isLoading: boolean
 }
 
 const initialState: IssuesState = {
   issues: [],
   currentPage: 1,
-  loadingState: 'notLoaded',
+  isLoading: false,
 }
 
-export const fetchIssues = createAsyncThunk(
+export const fetchIssues = createAsyncThunk<Issues, undefined, { state: RootState }>(
   'issues/fetchIssues',
-  async (currentPage: number) => {
-    const { data } = await github.getIssues(currentPage)
+  async (_, { getState }) => {
+    const { issues } = getState()
+    const { data } = await github.getIssues(issues.currentPage)
     return data
   },
 )
@@ -25,23 +28,20 @@ export const fetchIssues = createAsyncThunk(
 export const issueSlice = createSlice({
   name: 'issues',
   initialState,
-  reducers: {
-    toNextPage: (state) => {
-      state.currentPage += 1
-      state.loadingState = 'notLoaded'
-    },
-  },
+  reducers: {},
   extraReducers: (builder) => {
     builder.addCase(fetchIssues.fulfilled, (state, action) => {
       state.issues = [...state.issues, ...action.payload]
-      state.loadingState = 'loaded'
+      state.currentPage += 1
+      state.isLoading = false
     })
     builder.addCase(fetchIssues.pending, (state) => {
-      state.loadingState = 'loading'
+      state.isLoading = true
     })
   },
 })
 
-export const { toNextPage } = issueSlice.actions
+// TODO: 언젠가 사용하기 위한 보일러 플레이트
+// export const {} = issueSlice.actions
 
 export default issueSlice.reducer
